Refactor to replace separate sensor implimentation
