+++ make/source.mk
@@ -68,6 +68,7 @@ COMMON_SRC = \
             scheduler/scheduler.c \
             sensors/adcinternal.c \
             sensors/battery.c \
+            sensors/smartfuel.c \
             sensors/current.c \
             sensors/voltage.c \
             target/config_helper.c \
