+++ src/main/msp/msp_protocol_v2_betaflight.h
@@ -24,5 +24,3 @@
 #define MSP2_SEND_DSHOT_COMMAND             0x3003
 #define MSP2_GET_VTX_DEVICE_STATUS          0x3004
 #define MSP2_GET_OSD_WARNINGS               0x3005  // returns active OSD warning message text
-#define MSP2_GET_SMARTFUEL_CONFIG           0x3006
-#define MSP2_SET_SMARTFUEL_CONFIG           0x3007
