+++ src/main/sensors/battery.h
@@ -78,8 +78,9 @@ uint16_t getBatteryCurrentSample(void);
 uint16_t getLegacyBatteryCurrent(void);
 uint16_t getBatteryCapacity(void);
 uint32_t getBatteryCapacityUsed(void);
-int getBatterySmartFuel(void);
-int getBatterySmartConsumption(void);
+uint32_t getBatteryFuelConsumption(void);
+uint8_t getBatteryFuelLevel(void);
+bool batteryFuelLevelIsPercentage(void);
 
 uint8_t calculateBatteryPercentageRemaining(void);
