+++ src/main/telemetry/crsf.c
@@ -375,8 +375,8 @@ static void crsfFrameBatterySensor(sbuf_t *dst)
     sbufWriteU8(dst, CRSF_FRAMETYPE_BATTERY_SENSOR);
     sbufWriteU16BE(dst, getLegacyBatteryVoltage());
     sbufWriteU16BE(dst, getLegacyBatteryCurrent());
-    sbufWriteU24BE(dst, getBatteryCapacityUsed());
-    sbufWriteU8(dst, calculateBatteryPercentageRemaining());
+    sbufWriteU24BE(dst, getBatteryFuelConsumption());
+    sbufWriteU8(dst, getBatteryFuelLevel());
 }
 
 /*
@@ -733,10 +733,6 @@ static telemetrySensor_t crsfCustomTelemetrySensors[] =
     TLM_SENSOR(BATTERY_CURRENT,         0x1012,   200,  3000,    0,     U16),
     TLM_SENSOR(BATTERY_CONSUMPTION,     0x1013,   200,  3000,    0,     U16),
     TLM_SENSOR(BATTERY_CHARGE_LEVEL,    0x1014,   200,  3000,    0,     U8),
-#ifdef USE_SMARTFUEL
-    TLM_SENSOR(BATTERY_SMARTFUEL,       0x1015,   200,  3000,    0,     S8),
-    TLM_SENSOR(BATTERY_SMARTCONSUMPTION, 0x1016,  200,  3000,    0,     U16),
-#endif
 
     TLM_SENSOR(BATTERY_CELL_COUNT,      0x1020,   200,  3000,    0,     U8),
     TLM_SENSOR(BATTERY_CELL_VOLTAGE,    0x1021,   200,  3000,    0,     CellVolt),
