+++ src/main/telemetry/hott.c
@@ -291,7 +291,7 @@ static inline void hottEAMUpdateCurrentMeter(HOTT_EAM_MSG_t *hottEAMMessage)
 
 static inline void hottEAMUpdateBatteryDrawnCapacity(HOTT_EAM_MSG_t *hottEAMMessage)
 {
-    const uint16_t mAh = getBatteryCapacityUsed() / 10;
+    const uint16_t mAh = getBatteryFuelConsumption() / 10;
     hottEAMMessage->batt_cap_L = mAh & 0xFF;
     hottEAMMessage->batt_cap_H = mAh >> 8;
 }
