+++ src/main/telemetry/ltm.c
@@ -183,7 +183,7 @@ static void ltm_sframe(void)
         lt_statemode |= 2;
     ltm_initialise_packet('S');
     ltm_serialise_16(getBatteryVoltage() * 10); // vbat converted to mV
-    ltm_serialise_16((uint16_t)constrain(getBatteryCapacityUsed(), 0, UINT16_MAX)); // consumption in mAh (65535 mAh max)
+    ltm_serialise_16((uint16_t)constrain(getBatteryFuelConsumption(), 0, UINT16_MAX)); // consumption in mAh (65535 mAh max)
     ltm_serialise_8(constrain(scaleRange(getRssi(), 0, RSSI_MAX_VALUE, 0, 255), 0, 255));        // scaled RSSI (uchar)
     ltm_serialise_8(0);              // no airspeed
     ltm_serialise_8((lt_flightmode << 2) | lt_statemode);
