+++ src/main/telemetry/mavlink.c
@@ -138,7 +138,7 @@ static int16_t headingOrScaledMilliAmpereHoursDrawn(void)
 {
     if (isBatteryCurrentConfigured() && telemetryConfig()->mavlink_mah_as_heading_divisor > 0) {
         // In the Connex Prosight OSD, this goes between 0 and 999, so it will need to be scaled in that range.
-        return getBatteryCapacityUsed() / telemetryConfig()->mavlink_mah_as_heading_divisor;
+        return getBatteryFuelConsumption() / telemetryConfig()->mavlink_mah_as_heading_divisor;
     }
     // heading Current heading in degrees, in compass units (0..360, 0=north)
     return DECIDEGREES_TO_DEGREES(attitude.values.yaw);
@@ -227,7 +227,7 @@ void mavlinkSendSystemStatus(void)
     if (getBatteryState() < BATTERY_NOT_PRESENT) {
         batteryVoltage = isBatteryVoltageConfigured() ? getBatteryVoltage() * 10 : batteryVoltage;
         batteryAmperage = isBatteryCurrentConfigured() ? getBatteryCurrent() : batteryAmperage;
-        batteryRemaining = isBatteryVoltageConfigured() ? calculateBatteryPercentageRemaining() : batteryRemaining;
+        batteryRemaining = isBatteryVoltageConfigured() ? getBatteryFuelLevel() : batteryRemaining;
     }
 
     mavlink_msg_sys_status_pack(0, 200, &mavMsg,
