+++ src/main/telemetry/srxl.c
@@ -429,7 +429,7 @@ bool srxlFrameFlightPackCurrent(sbuf_t *dst, timeUs_t currentTimeUs)
   if ( isBatteryCurrentConfigured() ) {
 
     uint16_t amps = getLegacyBatteryCurrent();
-    uint16_t mah  = getBatteryCapacityUsed();
+    uint16_t mah  = getBatteryFuelConsumption();
     static uint16_t sentAmps;
     static uint16_t sentMah;
     static timeUs_t lastTimeSentFPmAh = 0;
